refactor: modularize extension functions
